requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
redis==5.0.1
aioredis==2.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
colorlog==6.7.0
loguru==0.7.2
marker-pdf
python-multipart